Remove duplicated code and add `calculatePercentagesFromResponse` function
