+++ GUI/src/pages/FeedbackPage.tsx
@@ -185,33 +185,8 @@ const FeedbackPage: React.FC = () => {
           return a;
         }, []);
 
-      const percentagesResponse = response.reduce(function (a: any, b: any) {
-        const res: any = {};
-        Object.keys(response[0]).forEach((e: string) => {
-          if (e != 'dateTime') {
-            res[e] = a[e] + b[e];
-          }
-        });
-        return res;
-      });
-
-      const percentages: any[] = [];
-      for (const key in percentagesResponse) {
-        const currentPercentage: any = {};
-        currentPercentage['name'] = key;
-        currentPercentage['value'] = parseFloat(
-          (
-            (percentagesResponse[key] /
-              Object.values(percentagesResponse).reduce<number>((a: any, b: any) => a + b, 0)) *
-            100
-          ).toFixed(1)
-        );
-        percentages.push(currentPercentage);
-      }
-
       chartData = {
         chartData: response,
-        percentagesData: percentages,
         colors: feedbackMetrics[0].subOptions!.map(({ id, color }) => {
           return {
             id: t(`feedback.plain_status_options.${id.toLowerCase()}`),
@@ -352,33 +327,8 @@ const FeedbackPage: React.FC = () => {
         return res;
       });
 
-      const percentagesResponse = chartResponse.reduce(function (a: any, b: any) {
-        const res: any = {};
-        Object.keys(chartResponse[0]).forEach((e: string) => {
-          if (e != 'dateTime') {
-            res[e] = a[e] + b[e];
-          }
-        });
-        return res;
-      });
-
-      const percentages: any[] = [];
-      for (const key in percentagesResponse) {
-        const currentPercentage: any = {};
-        currentPercentage['name'] = key;
-        currentPercentage['value'] = parseFloat(
-          (
-            (percentagesResponse[key] /
-              Object.values(percentagesResponse).reduce<number>((a: any, b: any) => a + b, 0)) *
-            100
-          ).toFixed(1)
-        );
-        percentages.push(currentPercentage);
-      }
-
       chartData = {
         chartData: chartResponse,
-        percentagesData: percentages,
         colors: feedbackMetrics[3].subOptions!.map(({ labelKey, color }) => {
           return {
             id: labelKey,
